requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
urllib3==1.26.9
yarl==1.7.2
